requirements.txt
scipy==1.4.1
tqdm==4.46.1
numpy==1.18.2
matplotlib==3.2.1
torchvision==0.9.0+cu111
torch==1.8.0+cu111
Pillow==8.3.2
